Coin_Clicker_Game
Coin Clicker. This is a basic game created in android studio by using kotlin language. When game starts the gold coins are falls in the top of the screen by clicking the falling coins earn some points, the game runs for several period of time, after the game time ends it shows the final points and highest score also.